requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.10
langchain-core>=0.1.10
python-dotenv>=1.0.0
openai>=1.0.0
pydantic
fastapi
uvicorn
python-multipart
gradio # or streamlit
requests
typing
psycopg2-binary